backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.9.1
boto3==1.28.66
botocore==1.31.66
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.3
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
drf-spectacular==0.26.5
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.0.1
pipreqs==0.4.13
platformdirs==3.11.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.4
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3
whitenoise==6.6.0
yarg==0.1.9
